+++ lib/src/Codec/EtcBlock4x4Encoding_RGBA8.cpp
@@ -276,7 +276,7 @@ namespace Etc
                 fMaxAlpha = fAlpha;
             }
             
-            srcAlpha[uiPixel] = fAlpha;
+            srcAlpha[uiPixel] = static_cast<uint8_t>(fAlpha);
         }
         
         assert(fMinAlpha >= 0);
@@ -368,13 +368,13 @@ namespace Etc
                     {
                         m_fError = (float)fBlockError;
 
-                        m_fBase = fBase;
-                        m_fMultiplier = fMultiplier;
-                        m_uiModifierTableIndex = uiTableEntry;
+                        m_fBase = static_cast<uint8_t>(fBase);
+                        m_fMultiplier = static_cast<uint8_t>(fMultiplier);
+                        m_uiModifierTableIndex = static_cast<uint8_t>(uiTableEntry);
                         
                         for (int uiPixel = 0; uiPixel < PIXELS; uiPixel++)
                         {
-                            m_auiAlphaSelectors[uiPixel] = auiBestSelectors[uiPixel];
+                            m_auiAlphaSelectors[uiPixel] = static_cast<uint8_t>(auiBestSelectors[uiPixel]);
                             
                             //m_afDecodedAlphas[uiPixel] = afBestDecodedAlphas[uiPixel] / 255.0f;
                         }
